refacto(API-1d) : move functions to school boundary
